requirements.txt
django
djangorestframework
djangorestframework_stubs
django-cors-headers
psycopg[binary]
pytest
pytest-django